add end and undend for project

+++ database/migrations/2024_01_29_032155_create_invoices_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('issued_date');
             $table->string('status');
             $table->date('due_date')->nullable();
-            $table->decimal('total');
+            $table->unsignedBigInteger('total');
             $table->string('invoice_pdf');
             $table->timestamps();
         });
